:recycle: Split OAuth2 routes into directory module by flow
Pure mechanical move of server/src/route/oauth2.rs into a directory module; no behavior change.
